Fix echo 0 hello, space in echo and multiple arguments for execve

+++ includes/exec.h
@@ -6,7 +6,7 @@
 /*   By: jlorette <jlorette@42angouleme.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/16 08:07:34 by jlorette          #+#    #+#             */
-/*   Updated: 2025/02/03 11:49:41 by jlorette         ###   ########.fr       */
+/*   Updated: 2025/02/03 16:05:57 by jlorette         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ void		exec(t_ast *ast, t_env *env_lst, int *flag_exit);
 char		*find_cmd(t_cmd *cmd, t_env *env, long *error);
 
 // execve
-size_t		cmd_size(t_cmd *cmd);
+int			cmd_size(t_cmd *cmd);
 void		fill_params_array(char **str, t_cmd *cmd, int *i);
 char		**join_params(t_cmd *cmd);
